+++ src/app/cluster/cluster-details/edit-cluster/edit-cluster.component.spec.ts
@@ -13,12 +13,14 @@ import {EditClusterComponent} from './edit-cluster.component';
 import {ClusterService} from '../../../core/services';
 import {fakeAWSDatacenter} from '../../../testing/fake-data/datacenter.fake';
 import {Subject} from 'rxjs';
+import {CoreModule} from '../../../core/core.module';
 
 const modules: any[] = [
   BrowserModule,
   BrowserAnimationsModule,
   SlimLoadingBarModule.forRoot(),
   SharedModule,
+  CoreModule,
 ];
 
 describe('EditClusterComponent', () => {
