+++ src/app/cluster/cluster-details/node-data-modal/node-data-modal.component.spec.ts
@@ -7,6 +7,7 @@ import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {ActivatedRoute, Router} from '@angular/router';
 import {SlimLoadingBarModule} from 'ng2-slim-loading-bar';
 
+import {CoreModule} from '../../../core/core.module';
 import {ApiService, ClusterService, DatacenterService, ProjectService, WizardService} from '../../../core/services';
 import {NodeDataService} from '../../../core/services/node-data/node-data.service';
 import {ClusterNameGenerator} from '../../../core/util/name-generator.service';
@@ -61,6 +62,7 @@ describe('NodeDataModalData', () => {
             NgReduxTestingModule,
             SharedModule,
             MatTabsModule,
+            CoreModule,
           ],
           declarations: [
             NodeDataModalComponent,
