+++ src/app/core/core.module.ts
@@ -20,8 +20,8 @@ import {NotificationComponent} from './components/notification/notification.comp
 import {ProjectSelectorComponent} from './components/sidenav/project/selector.component';
 import {SidenavComponent} from './components/sidenav/sidenav.component';
 import {AuthInterceptor, CheckTokenInterceptor, ErrorNotificationsInterceptor, LoaderInterceptor} from './interceptors';
-import {ApiService, Auth, AuthGuard, AuthzGuard, ClusterService, DatacenterService, LabelService, ParamsService} from './services';
-import {WizardService} from './services';
+import {ApiService, Auth, AuthGuard, AuthzGuard, ClusterService, DatacenterService, LabelService, ParamsService, WizardService} from './services';
+import {GlobalModule} from './services/global/global.module';
 import {NodeDataService} from './services/node-data/node-data.service';
 import {StepsService} from './services/wizard/steps.service';
 import {ClusterNameGenerator} from './util/name-generator.service';
@@ -32,6 +32,7 @@ const modules: any[] = [
   HttpClientModule,
   RouterModule,
   SharedModule,
+  GlobalModule,
   SimpleNotificationsModule.forRoot(),
 ];
