+++ src/app/project/edit-project/edit-project.component.html
@@ -18,7 +18,8 @@
         </mat-form-field>
       </form>
       <km-label-form title="Labels"
-                     [(labels)]="labels"></km-label-form>
+                     [(labels)]="labels"
+                     [asyncKeyValidators]=asyncLabelValidators></km-label-form>
     </div>
   </mat-dialog-content>
   <mat-dialog-actions>
