+++ src/app/wizard/set-cluster-spec/set-cluster-spec.component.html
@@ -107,7 +107,8 @@
         </div>
       </form>
       <km-label-form title="Labels"
-                     [(labels)]="labels"></km-label-form>
+                     [(labels)]="labels"
+                     [asyncKeyValidators]=asyncLabelValidators></km-label-form>
     </mat-card-content>
   </mat-card>
 </div>
